Reorganized project and added TS cuts for make-local-maps
