+++ simpledst-maps/src/private/make-random-local-maps.cc
@@ -29,10 +29,17 @@
 #include <iter-lhreco-proj/pickle.h>
 #include "TFile.h"
 #include "TTree.h"
+#include <TChain.h>
 
-//#include <photospline/splinetable.h>
-//#include <photospline/bspline.h>
-#include "esplines.h"
+#include "cuts.h"
+
+#include <astro/astro.h>
+#include <astro/time.h>
+#include <Direction.h>
+#include <solardipole.h>
+#include <SimpleTrigger.h>
+
+
 
 #ifdef HAWCNEST
 #include <hawcnest/Logging.h>
@@ -51,6 +58,8 @@ using boost::format;
 typedef Healpix_Map<double> SkyMap; 
 typedef boost::shared_ptr<SkyMap> SkyMapPtr; // Map shared pointer
 
+bool newConfig(string config);
+
 double 
 Sum(const SkyMap& map,unsigned int npix) 
 {
@@ -62,6 +71,25 @@ Sum(const SkyMap& map,unsigned int npix)
     return sumval;
 }
 
+bool filterCut(po::variables_map vm, SimpleDST dst) {
+
+  string filter = vm["filter"].as<string>();
+  string config = vm["config"].as<string>();
+  if (filter=="STA3" && dst.isSTA3)
+    return true;
+  if (config=="IT59" || config=="IT73" || config=="IT81") {
+    if ((filter=="STA8" && dst.isSTA8) ||
+        (filter=="NotSTA8" && dst.isSTA3 && !dst.isSTA8))
+      return true;
+  }
+  if (config=="IT81-2012" || config=="IT81-2013") {
+    if ((filter=="STA8" && dst.nStations>=8) ||
+        (filter=="NotSTA8" && dst.isSTA3 && dst.nStations<8))
+      return true;
+  }
+  return false;
+}
+
 
 
 int main(int argc, char* argv[])
@@ -70,14 +98,20 @@ int main(int argc, char* argv[])
 ////// Input ////////////////////////////////////////////////////////////////// 
 //*****************************************************************************
     std::string foldername;
+    std::string method("sid");
+    std::string config;
+    std::string filter;
     unsigned int nTimesteps;
     unsigned int nIterations;
     unsigned int nSectors;
     int nsideOut;
     int nmaps;
+    bool sundp;
+    bool show_progress;
     double lon;
     double lat;
     double elogmin, elogmax, rloglmax,ldirmin,ndirmin;
+    double slogmin, slogmax;
     std::vector< std::string > input;
     std::string coords = "Azimuth/ Zenith";
     int nEvents = 0;
@@ -96,11 +130,17 @@ int main(int argc, char* argv[])
              ("spline", po::value<string>(), "File containing spline tables")
              ("elogmin", po::value<double>(&elogmin)->default_value(0), "Minimum energy")
              ("elogmax", po::value<double>(&elogmax)->default_value(0), "Maximum energy (0: no cut)")
+             ("slogmin", po::value<double>(&slogmin)->default_value(0), "Minimum logS125")
+             ("slogmax", po::value<double>(&slogmax)->default_value(0), "Maximum logS125 (0: no cut)")
              ("ldirc_min", po::value<double>(&ldirmin)->default_value(0), "Minimum ldir (0: no cut)")
              ("ndirc_min", po::value<double>(&ndirmin)->default_value(0), "Minimum ndir (0: no cut)")
              ("rloglmax", po::value<double>(&rloglmax)->default_value(0), "Maximum rlogL (0: no cut)")
              ("nsideout", po::value<int>(&nsideOut)->default_value(256), "Healpix Nside for output map") 
              ("nmaps", po::value<int>(&nmaps)->default_value(10), "Number of random maps togenerate") 
+             ("sundp", po::bool_switch(&sundp)->default_value(false), "subtract solar dipole")
+             ("config", po::value<string>(&config), "Detector configuration") 
+             ("method", po::value<string>(&method), "Sidereal, Anti, Solar, Extended")
+             ("filter", po::value<string>(&filter), "Filter for IceTop data")
              ;
      
         //po::store(po::parse_command_line(argc, argv, desc),  vm); // can throw 
@@ -133,25 +173,39 @@ int main(int argc, char* argv[])
 //*****************************************************************************
 ////// Initialize ///////////////////////////////////////////////////////////// 
 //*****************************************************************************
+    const char* masterTree;
+    const char* triggerTree;
+    string detector = config.substr(0,2); 
+    if (detector == "IC") { 
+        masterTree = "CutDST"; 
+        triggerTree = "TDSTTriggers"; 
+    } 
+    if (detector == "IT") { 
+        masterTree = "master_tree"; 
+        triggerTree = "";   // Unused? Will probably break IT functionality...  
+    }
+
 
     // TFile f("Event.root");
-    TFile file(input[0].c_str());
+    TChain cutDST(masterTree);
+    for (unsigned i = 0; i < input.size(); ++i) { 
+        cutDST.Add(input[i].c_str()); 
+    } 
+    SimpleDST dst(&cutDST, config);
 
-    TTree* tree = (TTree*) file.Get("CutDST");
-    float LLHZenithDeg, LLHAzimuthDeg, RADeg, DecDeg, RLogL;
-    double ModJulDay, LocalMST, rndMJD;
-    UShort_t NChannels; 
-    UInt_t NDirHits, LDir;
+    TChain trigDST(triggerTree);
+    if (newConfig(config)) { 
+        for (unsigned i = 0; i < input.size(); ++i) { 
+            trigDST.Add(input[i].c_str()); 
+        }
+    } 
+    SimpleTrigger dst_trig(&trigDST);
 
-    tree->SetBranchAddress("ModJulDay", &ModJulDay);
-    tree->SetBranchAddress("LocalMST", &LocalMST);
-    tree->SetBranchAddress("LLHZenithDeg", &LLHZenithDeg);
-    tree->SetBranchAddress("LLHAzimuthDeg", &LLHAzimuthDeg);
-    tree->SetBranchAddress("RADeg", &RADeg);
-    tree->SetBranchAddress("DecDeg", &DecDeg);
-    tree->SetBranchAddress("NChannels", &NChannels);
-    tree->SetBranchAddress("NDirHits", &NDirHits);
-    tree->SetBranchAddress("LDir", &LDir);
+    cout << "Number of chained files: " << cutDST.GetNtrees() << endl; 
+    Long64_t nEntries = cutDST.GetEntries();
+
+    double rndMJD;
+
 
     // Read in spline tables if provided
     photospline::splinetable<> spline;
@@ -199,7 +253,9 @@ int main(int argc, char* argv[])
                             log_info("....successfully created !");
             }
 
-            float theta, phi;
+            double eventweight;
+            double weightsum = 0.;
+            float zenith, azimuth;
             int timeBin;
             int prevTimeBin = 0 ;
             int firstTimeBin = 0 ;
@@ -209,40 +265,66 @@ int main(int argc, char* argv[])
             bool init=false;
             SkyMapPtr tmp_map;
 
-            for (int i = 0, N = tree->GetEntries(); i < N; ++i) {
-               tree->GetEntry(i); 
+            for (int i = 0; i < nEntries; ++i) {
+               cutDST.GetEntry(i); 
+               if (newConfig(config)) { 
+                   trigDST.GetEntry(i); 
+               }
 
-               if ( (rloglmax > 0 ) && (RLogL > rloglmax) )
-                  continue;
+               if ( (rloglmax > 0 ) && (dst.RLogL > rloglmax) )
+                  continue; 
 
-               
-               //std::cout << "NDirHits=" << NDirHits << " ndirmin = " << ndirmin << std::endl;
-               //if ( NDirHits < ndirmin ) 
-               if ( NDirHits < ndirmin*cos(LLHZenithDeg) ) 
-                  continue;
+               double mjd = dst.ModJulDay; 
+               double lst = GetGMST(mjd); 
+               if (method == "anti") { 
+                   lst = GetGMAST(mjd); 
+               } else if (method == "ext") { 
+                   lst = GetGMEST(mjd); 
+               } else if (method == "solar") { 
+                   lst= ( mjd - int(mjd) )* 24.; 
+               }
 
-               //std::cout << "LDir= " << LDir << " ldirmin = " << ldirmin << std::endl;
-               //if ( LDir < ldirmin ) 
-               if ( LDir < ldirmin*cos(LLHZenithDeg) ) 
-                  continue;
-
-               currMJD = ModJulDay;
                // https://en.wikipedia.org/wiki/Time_in_Antarctica
-               //
-               //rndMJD = histMJD[mEntry]->GetRandom();
+               azimuth = dst.LLHAzimuthDeg/180.*M_PI; 
+               //phi = RADeg/180.*M_PI; 
+               zenith = dst.LLHZenithDeg/180.*M_PI; 
+               //theta = DecDeg/180.*M_PI; 
+
+
+               if ( dst.NDirHits < ndirmin*cos(zenith) ) 
+                  continue;
+
+               if ( dst.LDir < ldirmin*cos(zenith) ) 
+                  continue;
+
                double rngMST = rand.Uniform(24.0);
                
                timeBin = int( nTimeBins*rngMST/24.0 ); 
 
-               phi = LLHAzimuthDeg/180.*M_PI; 
-               //phi = RADeg/180.*M_PI; 
-               theta = LLHZenithDeg/180.*M_PI; 
-               //theta = DecDeg/180.*M_PI; 
+               // Reconstruction cuts
+               if (!dst.isReco || zenith != zenith || azimuth != azimuth) 
+                   continue;
+
+               // IceTop filter cut
+               if (vm.count("filter") && !filterCut(vm, dst)) 
+                   continue;
                
-               if ((vm.count("spline")) && !ICenergyCut(NChannels, spline, theta, elogmin, elogmax))
+               if (vm.count("spline") && !ICenergyCut(dst, spline, zenith, elogmin, elogmax))
                   continue;
 
-               pointing pt(theta, phi);
+               // Energy cuts for IceTop and IceCube
+               if (detector == "IT" && !ITenergyCut(dst, elogmin,elogmax)) 
+                   continue;
+
+               pointing pt(zenith, azimuth);
+
+               eventweight = 1.0;
+               if (sundp) { 
+                   Direction dir(zenith,azimuth); 
+                   Equatorial eq = GetEquatorialFromDirection(dir, mjd); 
+                   eventweight = solar_dipole(mjd, eq.ra, eq.dec,true); 
+               } 
+               weightsum += eventweight;
 
                tmp_map = Nmap[timeBin % 360];// make sure tsid < 360 deg
                int j = tmp_map->ang2pix(pt);
@@ -250,6 +332,7 @@ int main(int argc, char* argv[])
                nEvents++;
 
             }
+            std::cout << std::endl;
 
 
             for (unsigned int degbin=0;degbin< 360;degbin++) 
@@ -269,8 +352,21 @@ int main(int argc, char* argv[])
 
                  SkyMap jmap = *(Nmap[degbin]);
                  write_Healpix_map_to_fits(fitsOut, jmap, PLANCK_FLOAT64);
-                 //write_Healpix_map_to_fits(fitsOut, jmap, FITSUTIL<double>::DTYPE);
                  fitsOut.close();
             }
+            std::cout << "passing rate: " << nEvents << "/" << nEntries << std::endl;
+            std::cout << "weightsum: " << weightsum << std::endl;
+            
     }
 }
+
+bool newConfig(string config) {
+
+  if (config=="IC86-2011" || config=="IC86-2012" || config=="IC86-2013" || 
+      config=="IC86-2014" || config=="IC86-2015") {
+    return false;
+  }
+  return true;
+}
+
+
