+++ simpledst-maps/src/scripts/submit_all.py.in
@@ -26,7 +26,7 @@ if __name__ == "__main__":
             default='/data/user/fmcnally/anisotropy/sim/IC86_20904_hist_spline.fits', 
             help='Path to spline files')
     p.add_argument('--cvmfs', dest='cvmfs',
-            default='/cvmfs/icecube.opensciencegrid.org/py3-v4.2.1/',
+            default='@CVMFS_SROOTBASE@',
             help='CVMFS environment')
     p.add_argument('--submit_dir', dest='submit_dir',
             default='@CRA_BUILD@/submit-dir',
